(feature) CREATE DELETE MODAL FOR CUSTOMER
I first started by creating the modal and linking it with the ts file.
I used the selectedClient variable the record to be deleted and sent the
id to the server through a delete http request. I procceded to create a
new api end point to recieve the new delete request.